// packages/lego-webapp/redux/models/User.ts
export interface AchievementData {
  name: string;
  description: string;
  rarity: number;
  hidden: boolean;
}